requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3
openslide-python==1.3.1
Pillow==10.4.0
python-multipart==0.0.9
SQLAlchemy==2.0.30
aiosqlite==0.20.0
Jinja2==3.1.4